write benchmarks for bcrypt, scrypt, mpw
