+++ src/benchmark.rs
@@ -15,25 +15,68 @@
  * along with Master Password. If not, see <http://www.gnu.org/licenses/>.
  */
 
-use std::time;
-use ring::{digest, hmac};
-use bcrypt::hash;
+// use std::time;
+// use ring::{digest, hmac};
+// use bcrypt::hash;
+// use crate::common::SiteType;
+// use crate::common::SiteVariant;
+// use crate::core::master_key_for_user;
+// use crate::core::password_for_site;
+//
+// fn calc_speed(elapsed: time::Duration, iterations: u32) -> f64 {
+//     let seconds = (elapsed.as_secs() as f64) + (elapsed.subsec_nanos() as f64 / 1000_000_000.0);
+//     let speed = iterations as f64 / seconds;
+//     println!(" * Completed {} iterations in {:0.2} seconds at {:0.2} ops/s.",
+//              iterations,
+//              seconds,
+//              speed);
+//     speed
+// }
+
 use crate::common::SiteType;
 use crate::common::SiteVariant;
 use crate::core::master_key_for_user;
 use crate::core::password_for_site;
+use bcrypt::hash;
+use ring::{digest, hmac};
+extern crate test as TEST;
+use TEST::Bencher;
 
-fn calc_speed(elapsed: time::Duration, iterations: u32) -> f64 {
-    let seconds = (elapsed.as_secs() as f64) + (elapsed.subsec_nanos() as f64 / 1000_000_000.0);
-    let speed = iterations as f64 / seconds;
-    println!(" * Completed {} iterations in {:0.2} seconds at {:0.2} ops/s.",
-             iterations,
-             seconds,
-             speed);
-    speed
+#[bench]
+fn hmac_sha256(b: &mut Bencher) {
+    let full_name = "Robert Lee Mitchel";
+    let master_password = "banana colored duckling";
+    let site_variant = SiteVariant::Password;
+    let algo = "3";
+    let master_key: [u8; 64] =
+        master_key_for_user(&full_name, &master_password, &algo, &site_variant).unwrap();
+    b.iter(|| {
+        hmac::sign(
+            &hmac::SigningKey::new(&digest::SHA256, &master_key),
+            "".as_bytes(),
+        );
+    })
 }
 
-pub fn mpw_bench() {
+#[bench]
+fn bcrypt9(b: &mut Bencher) {
+    let master_password = "banana colored duckling";
+    let bcrypt_cost = 9;
+    b.iter(|| hash(master_password, bcrypt_cost))
+}
+
+#[bench]
+fn scrypt_mpw(b: &mut Bencher) {
+    let full_name = "Robert Lee Mitchel";
+    let master_password = "banana colored duckling";
+    let site_variant = SiteVariant::Password;
+    let algo = "3";
+
+    b.iter(|| master_key_for_user(full_name, master_password, algo, &site_variant));
+}
+
+#[bench]
+fn mpw(b: &mut Bencher) {
     let full_name = "Robert Lee Mitchel";
     let master_password = "banana colored duckling";
     let site_name = "masterpasswordapp.com";
@@ -43,62 +86,16 @@ pub fn mpw_bench() {
     let site_context = "";
     let algo = "3";
 
-    println!("\n## Benchmarking Master Password\n");
-
-    let master_key: [u8; 64] = master_key_for_user(&full_name, &master_password, &algo, &site_variant)
-        .unwrap();
-    let iterations = 3_000_000;
-    println!("### Performing {} iterations of {}:",
-             iterations,
-             "hmac-sha-256");
-    let start = time::Instant::now();
-    for _ in 1..iterations {
-        hmac::sign(&hmac::SigningKey::new(&digest::SHA256, &master_key),
-                   "".as_bytes());
-    }
-    let hmac_sha256_speed = calc_speed(start.elapsed(), iterations);
-
-    let bcrypt_cost = 9;
-    let iterations = 1000;
-    println!("### Performing {} iterations of {}:", iterations, "bcrypt9");
-    let start = time::Instant::now();
-    for _ in 1..iterations {
-        let _ = hash(master_password, bcrypt_cost);
-    }
-    let bcrypt_9_speed = calc_speed(start.elapsed(), iterations);
-
-    let iterations = 50;
-    println!("### Performing {} iterations of {}:",
-             iterations,
-             "scrypt_mpw");
-    let start = time::Instant::now();
-    for _ in 1..iterations {
-        master_key_for_user(full_name, master_password, algo, &site_variant);
-    }
-    let scrypt_speed = calc_speed(start.elapsed(), iterations);
-
-    let iterations = 50;
-    println!("### Performing {} iterations of {}:", iterations, "mpw");
-    let start = time::Instant::now();
-    for _ in 1..iterations {
+    b.iter(|| {
         let key = master_key_for_user(full_name, master_password, algo, &site_variant).unwrap();
-        password_for_site(&key,
-                          site_name,
-                          &site_type,
-                          &site_counter,
-                          &site_variant,
-                          &site_context,
-                          algo);
-    }
-    let mpw_speed = calc_speed(start.elapsed(), iterations);
-
-    println!("\n### Summary for this machine:");
-    println!(" * mpw is {:0.2} times slower than hmac-sha-256.",
-             hmac_sha256_speed / mpw_speed);
-    println!(" * mpw is {:0.2} times slower than bcrypt (cost 9).",
-             bcrypt_9_speed / mpw_speed);
-    println!(" * scrypt is {:0.2} times slower than bcrypt (cost 9).",
-             bcrypt_9_speed / scrypt_speed);
-
-    println!("\n## Benchmark complete");
+        password_for_site(
+            &key,
+            site_name,
+            &site_type,
+            &site_counter,
+            &site_variant,
+            &site_context,
+            algo,
+        );
+    });
 }
