+++ src/common.rs
@@ -214,49 +214,3 @@ mod tests {
 }
 
 
-mod benches {
-    use ring::{digest, hmac};
-    use bcrypt::hash;
-    use crate::common::SiteType;
-    use crate::common::SiteVariant;
-    use crate::core::master_key_for_user;
-    use crate::core::password_for_site;
-    extern crate test as TEST;
-    use TEST::Bencher;
-
-    #[bench]
-    fn mpw_bench(b: &mut Bencher) {
-        let full_name = "Robert Lee Mitchel";
-        let master_password = "banana colored duckling";
-        let site_name = "masterpasswordapp.com";
-        let site_counter = 1_i32;
-        let site_type = SiteType::Long;
-        let site_variant = SiteVariant::Password;
-        let site_context = "";
-        let algo = "3";
-
-        b.iter(|| {
-            let master_key: [u8; 64] = master_key_for_user(&full_name, &master_password, &algo, &site_variant)
-                .unwrap();
-            });
-    }
-
-    #[bench]
-    fn hmac_sha256(b: &mut Bencher) {
-        let full_name = "Robert Lee Mitchel";
-        let master_password = "banana colored duckling";
-        let site_name = "masterpasswordapp.com";
-        let site_counter = 1_i32;
-        let site_type = SiteType::Long;
-        let site_variant = SiteVariant::Password;
-        let site_context = "";
-        let algo = "3";
-        let master_key: [u8; 64] = master_key_for_user(&full_name, &master_password, &algo, &site_variant)
-            .unwrap();
-        b.iter(|| {
-            hmac::sign(&hmac::SigningKey::new(&digest::SHA256, &master_key),
-            "".as_bytes());
-        })
-    }
-
-}
